requirements.txt
amqp==2.5.2
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coverage==5.1
cryptography==2.9
dnspython==1.16.0
docker==4.2.0
email-validator==1.0.5
fastapi==0.54.1
gevent==1.5.0
greenlet==0.4.15
h11==0.9.0
httptools==0.1.1
idna==2.9
kombu==4.6.8
mock==4.0.2
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
starlette==0.13.2
supervisor==4.1.0
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.57.0
websockets==8.1
